Begins the creation of a naive solution for the assignment
Adds a player character with input for movement and attack. It fires projectiles which gets destroyed once they are outside of the border. The projectiles are created and handled by the ProjectileManager.
Also adds an EnemyManager which can create one type of enemy. Currently the enemy just walks to the players position.