Adds the last things for the naive implementation
Adds a timer and timerManager which makes it easier to create the timers for enemy/player attack and enemy spawner.
Adds a new enemy which is a wizard that shoots purple fireballs(arcaneballs) toawrds the player.
Adds enum class DamageType to projectiles so the arcaneballs damages the player and the fireballs damages the enemies.