Create a pool for enemies and projectiles
Creates pool enemies and projectiles during startup time and so they are not created in runtime. Uses a vector the holds the active elements and calls on their update and render, uses another vector that holds the inactive elements and is simply just a container that exists for the active element vector to get elements from.