Changes the way to spawn/remove enemies and projectiles. Changes instances to smart pointers if doable
Now enemies and projectiles are inserted in order based on IDs, from lowest to highest.
Adds a binary search function that locates the object based on the ID.

Changes instances to smart pointers if doable. Enemies and projectiles had to be ordinary pointers since it gave template errors when attempting to use it as a template.
The part that is from SDL like sprites, texts, window and renderer remains ordinary pointers as well.